feat: add donor dashboard and registration components
- Implemented a registration component with a comprehensive form and styles.
- Added home component with layout for displaying information and actions.
