+++ src/app/app.routes.ts
@@ -1,11 +1,14 @@
 import { Routes } from '@angular/router';
-import { Home } from './components/pages/home/home';
+import { Home } from './components/public-pages/home/home';
 import { Login } from './shared/components/login/login';
+import { Register } from './components/donor/register/register';
 
 export const routes: Routes = [
   { path: '', redirectTo: '/home', pathMatch: 'full' },
   { path: 'home', component: Home },
   { path: 'login', component: Login },
+  { path: 'donor/register', component: Register },
+
 
   // { path: 'about', component: SobreNosComponent },
   // { path: '**', component: NotFoundComponent }
