+++ src/app/components/donor/register/register.css
@@ -35,3 +35,7 @@
   border-radius: 8px;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
+
+a {
+  cursor: pointer;
+}
