--- src/app/components/pages/home/home.css
@@ -1,113 +1,0 @@
-/* ===== HERO ===== */
-.hero {
-  background: #fff;
-  padding: 40px 20px;
-  text-align: center;
-}
-
-.hero-content {
-  max-width: 800px;
-  margin: auto;
-}
-
-.hero-image {
-  max-width: 200px;
-  margin-bottom: 20px;
-}
-
-.text h1 {
-  font-size: 1.6rem;
-  font-weight: bold;
-  margin-bottom: 10px;
-  color: #333;
-}
-
-.text p {
-  font-size: 1rem;
-  color: #555;
-  margin-bottom: 20px;
-}
-
-.buttons {
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-}
-
-.buttons .btn {
-  padding: 10px 20px;
-  border-radius: 6px;
-  border: none;
-  font-weight: 500;
-  cursor: pointer;
-}
-
-.btn.institution {
-  background: #e63946;
-  color: #fff;
-}
-
-.btn.donor {
- background: #f3e7e8;
-  border: 1px solid #e63946;
-  color: #1b0e0e;;
-}
-
-/* ===== MAPA ===== */
-.map-section {
-  background: #e63946;
-  color: #fff;
-  text-align: center;
-  padding: 15px 0;
-}
-
-.map-section h2 {
-  margin: 0;
-  font-size: 1.2rem;
-  font-weight: bold;
-}
-
-.map-container img {
-  width: 100%;
-  display: block;
-}
-
-/* ===== DOAÇÕES ===== */
-.donations-section {
-  background: #e63946;
-  color: #fff;
-  text-align: center;
-  padding: 15px 0;
-}
-
-.donations-section h2 {
-  margin: 0;
-  font-size: 1.2rem;
-  font-weight: bold;
-}
-
-.donation-box {
-  background: #f8f9fa;
-  margin: 0 auto;
-  padding: 20px;
-  max-width: 500px;
-  text-align: center;
-  color: #000;
-}
-
-.donation-box p {
-  margin: 0;
-  font-size: 1rem;
-  color: #555;
-}
-
-.donation-box h3 {
-  margin: 10px 0;
-  font-size: 2rem;
-  font-weight: bold;
-}
-
-.donation-box .positive {
-  color: #28a745;
-  font-weight: bold;
-}
