--- src/app/components/pages/home/home.html
@@ -1,41 +1,0 @@
-<app-header></app-header>
-
-<!-- HERO (conteúdo principal da home) -->
-<section class="hero">
-  <div class="hero-content">
-    <img src="../../../../assets/corcaoProjeto.png" alt="Coração solidário" class="hero-image">
-    <div class="text">
-      <h1>Conectando solidariedade e tecnologia</h1>
-      <p>
-        Facilitamos doações seguras e transparentes,<br>
-        conectando doadores e instituições verificadas.
-      </p>
-      <div class="buttons">
-        <button class="btn institution">Sou Instituição</button>
-        <button class="btn donor">Sou Doador</button>
-      </div>
-    </div>
-  </div>
-</section>
-
-<!-- MAPA -->
-<section class="map-section">
-  <h2>Instituições Próximas</h2>
-  <div class="map-container">
-    <img src="#pesquisarsobre maps" alt="Mapa de instituições próximas">
-  </div>
-</section>
-
-<!-- DOAÇÕES -->
-<section class="donations-section">
-  <h2>Doações em Tempo Real</h2>
-  <div class="donation-box">
-    <p>Total Doado</p>
-    <h3>R$ 1.234.567</h3>
-    <span class="positive">+15%</span>
-  </div>
-</section>
-
-
-
-<app-footer></app-footer>
