+++ src/app/components/public-pages/home/home.html
@@ -1,0 +1,57 @@
+<app-header></app-header>
+
+  <!--Cabeçalho-->
+        <div class="px-40 flex flex-1 justify-center py-5">
+          <div class="layout-content-container flex flex-col max-w-[960px] flex-1">
+            <div class="@container">
+              <div class="@[480px]:p-4">
+                <div
+                  class="flex min-h-[480px] flex-col gap-6 bg-cover bg-center bg-no-repeat @[480px]:gap-8 @[480px]:rounded-lg items-center justify-center p-4"
+                  style='background-image: linear-gradient(rgba(0, 0, 0, 0.1) 0%, rgba(0, 0, 0, 0.4) 100%), url("corcaoProjeto.png");'
+                >
+                  <div class="flex flex-col gap-2 text-center">
+                    <h1
+                      class="text-white text-4xl font-black leading-tight tracking-[-0.033em] @[480px]:text-5xl @[480px]:font-black @[480px]:leading-tight @[480px]:tracking-[-0.033em]"
+                    >
+                      Conectando solidariedade e tecnologia
+                    </h1>
+                    <h2 class="text-white text-sm font-normal leading-normal @[480px]:text-base @[480px]:font-normal @[480px]:leading-normal">
+                      Facilitamos doações seguras e transparentes, conectando doadores e instituições verificadas.
+                    </h2>
+                  </div>
+                  <div class="flex-wrap gap-3 flex justify-center">
+                    <!--Botões-->
+                    <button
+                      class="flex min-w-[84px] max-w-[480px] cursor-pointer items-center justify-center overflow-hidden rounded-lg h-10 px-4 @[480px]:h-12 @[480px]:px-5 bg-[#ea2832] text-[#fcf8f8] text-sm font-bold leading-normal tracking-[0.015em] @[480px]:text-base @[480px]:font-bold @[480px]:leading-normal @[480px]:tracking-[0.015em]"
+                    >
+                      <span class="truncate"><a href="login">Sou Instituição</a></span>
+                    </button>
+                    <button
+                      class="flex min-w-[84px] max-w-[480px] cursor-pointer items-center justify-center overflow-hidden rounded-lg h-10 px-4 @[480px]:h-12 @[480px]:px-5 bg-[#f3e7e8] text-[#1b0e0e] text-sm font-bold leading-normal tracking-[0.015em] @[480px]:text-base @[480px]:font-bold @[480px]:leading-normal @[480px]:tracking-[0.015em]"
+                    >
+                      <span class="truncate"><a href="login">Sou Doador</a></span>
+                    </button>
+                  </div>
+                </div>
+              </div>
+            </div>
+            <h2 class="text-[#1b0e0e] text-[22px] font-bold leading-tight tracking-[-0.015em] px-4 pb-3 pt-5">Instituições Próximas</h2>
+            <div class="flex px-4 py-3">
+              <div
+                class="w-full bg-center bg-no-repeat aspect-video bg-cover rounded-lg object-cover"
+                style='background-image: url("https://lh3.googleusercontent.com/aida-public/AB6AXuD4F27YgelHNhhrP0N5_o2rU02JNJAxu8HAFd38bnZC3Ix6bXyiBpCvh9tXN2eaqsCDkPUrH7fmG5iKK4XPT0CAv6Nc_DB1ymTYZKRKL9NpJvtGdSlg5-9tX-H0ypmHM6t4OyiVANeqrCtqgoZaNPLRHW3BaYX730nWEFnvIpOf_4q64k_vxcV_ShV1CJkrH-qxOU8o7APTMMLZT0Mc88oDMOx7wSSuzDZJSokT27iun7lQqwjxnnPjGAbzAd2hc8KCFueT_nAcKvI");'
+              ></div>
+            </div>
+            <h2 class="text-[#1b0e0e] text-[22px] font-bold leading-tight tracking-[-0.015em] px-4 pb-3 pt-5">Doações em Tempo Real</h2>
+            <div class="flex flex-wrap gap-4 p-4">
+              <div class="flex min-w-[158px] flex-1 flex-col gap-2 rounded-lg p-6 bg-[#f3e7e8]">
+                <p class="text-[#1b0e0e] text-base font-medium leading-normal">Total Doado</p>
+                <p class="text-[#1b0e0e] tracking-light text-2xl font-bold leading-tight">R$ 1.234.567</p>
+                <p class="text-[#07885d] text-base font-medium leading-normal">+15%</p>
+              </div>
+            </div>
+          </div>
+        </div>
+
+
+<app-footer></app-footer>
