--- src/app/components/user/register/register.html
@@ -1,1 +1,0 @@
-<p>register works!</p>
