--- src/app/components/user/register/register.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-register',
-  imports: [],
-  templateUrl: './register.html',
-  styleUrl: './register.css'
-})
-export class Register {
-
-}
