+++ src/app/core/services/Donator.ts
@@ -3,6 +3,6 @@ import { Injectable } from '@angular/core';
 @Injectable({
   providedIn: 'root'
 })
-export class User {
+export class Donator {
   
 }
