+++ src/app/shared/components/footer/footer.html
@@ -5,6 +5,7 @@
         <h3>Conecta Doa</h3>
       </div>
 
+      <!-- Links do rodapé -->
       <div class="footer-section">
         <a routerLink="/about">Sobre Nós</a>
         <a routerLink="/contato">Contato</a>
@@ -12,6 +13,7 @@
         <a routerLink="/privacity">Política de Privacidade</a>
       </div>
 
+      <!-- Direitos Autorais -->
       <div class="footer-bottom">
         <p>&copy; {{ anoAtual }} Conecta Doa. Todos os direitos reservados.</p>
       </div>
