+++ src/app/shared/components/footer/footer.ts
@@ -7,7 +7,7 @@ import { RouterModule } from '@angular/router';
   standalone: true,
   imports: [CommonModule, RouterModule],
   templateUrl: './footer.html',
-  styleUrls: ['./footer.css'] 
+  styleUrls: ['./footer.css']
 })
 export class FooterComponent {
   anoAtual: number = new Date().getFullYear();
