+++ src/app/shared/components/header/header.css
@@ -12,6 +12,14 @@
   display: flex;
   align-items: center;
   gap: 10px;
+  max-width: 150px;
+}
+img {
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+  object-fit: cover;
+  width: 500px, ;
 }
 
 .menu {
