+++ src/app/shared/components/header/header.html
@@ -1,6 +1,6 @@
 <header class="header">
   <div class="logo">
-    <img src="../assets/conectaDoacLogo.jpg" alt="Logo ConectaDoa">
+    <img src="/conectaDoacLogo.jpg" alt="Logo ConectaDoa">
     <h2>ConectaDoa</h2>
   </div>
 
@@ -12,7 +12,7 @@
 
     <div class="auth">
       <button class="btn vermelho" routerLink="/login">Entrar</button>
-      <button class="btn claro" routerLink="/register">Cadastrar</button>
+      <button class="btn claro" routerLink="/donor/register">Cadastrar</button>
     </div>
   </nav>
