+++ src/styles.css
@@ -1,1 +1,16 @@
-/* You can add global styles to this file, and also import other style files */
+:root {
+      --primary-color: #ea2a33;
+      --background-color: #fcf8f8;
+      --text-primary: #1b0e0e;
+      --text-secondary: #994d51;
+      --input-bg: #f3e7e8;
+      --input-placeholder: #994d51;
+    }
+    body {
+      font-family: 'Plus Jakarta Sans', sans-serif;
+    }
+
+    a {
+      text-decoration: none;
+      cursor: pointer;
+    }
